moved multiOutbound to internal package
