+++ crossdock/server/yarpc/phone.go
@@ -28,6 +28,7 @@ import (
 
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/encoding/json"
+	"go.uber.org/yarpc/internal/channel"
 	"go.uber.org/yarpc/transport"
 	ht "go.uber.org/yarpc/transport/http"
 	tch "go.uber.org/yarpc/transport/tchannel"
@@ -95,7 +96,7 @@ func Phone(ctx context.Context, reqMeta yarpc.ReqMeta, body *PhoneRequest) (*Pho
 	defer outbound.Stop()
 
 	// TODO use reqMeta.Service for caller
-	client := json.New(transport.MultiOutboundChannel("yarpc-test", body.Service, transport.Outbounds{
+	client := json.New(channel.MultiOutbound("yarpc-test", body.Service, transport.Outbounds{
 		Unary: outbound,
 	}))
 	resBody := PhoneResponse{
