+++ encoding/raw/outbound_test.go
@@ -27,6 +27,7 @@ import (
 	"testing"
 
 	"go.uber.org/yarpc"
+	"go.uber.org/yarpc/internal/channel"
 	"go.uber.org/yarpc/transport"
 	"go.uber.org/yarpc/transport/transporttest"
 
@@ -78,7 +79,7 @@ func TestCall(t *testing.T) {
 
 	for _, tt := range tests {
 		outbound := transporttest.NewMockUnaryOutbound(mockCtrl)
-		client := New(transport.MultiOutboundChannel(caller, service,
+		client := New(channel.MultiOutbound(caller, service,
 			transport.Outbounds{
 				Unary: outbound,
 			}))
