+++ encoding/thrift/outbound_test.go
@@ -28,6 +28,7 @@ import (
 	"testing"
 	"time"
 
+	"go.uber.org/yarpc/internal/channel"
 	"go.uber.org/yarpc/transport"
 	"go.uber.org/yarpc/transport/transporttest"
 
@@ -183,7 +184,7 @@ func TestClient(t *testing.T) {
 		opts = append(opts, Protocol(proto))
 		c := New(Config{
 			Service: "MyService",
-			Channel: transport.MultiOutboundChannel("caller", "service",
+			Channel: channel.MultiOutbound("caller", "service",
 				transport.Outbounds{
 					Unary: trans,
 				}),
