+++ inject_test.go
@@ -8,6 +8,7 @@ import (
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/encoding/json"
 	"go.uber.org/yarpc/encoding/raw"
+	"go.uber.org/yarpc/internal/channel"
 	"go.uber.org/yarpc/transport"
 	"go.uber.org/yarpc/transport/transporttest"
 
@@ -124,7 +125,7 @@ func TestInjectClientSuccess(t *testing.T) {
 			target: &struct {
 				Client json.Client `service:"foo"`
 			}{
-				Client: json.New(transport.MultiOutboundChannel(
+				Client: json.New(channel.MultiOutbound(
 					"foo",
 					"bar",
 					transport.Outbounds{
